# db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Room struct {
	ID   int32  `json:"id"`
	Name string `json:"name"`
}

type User struct {
	ID        int32            `json:"id"`
	Username  string           `json:"username"`
	Email     string           `json:"email"`
	Password  string           `json:"password"`
	Age       pgtype.Int4      `json:"age"`
	RoomID    pgtype.Int4      `json:"room_id"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}